GLOBAL: Unify input, r_screen
