+++ Makefile.nx
@@ -27,7 +27,8 @@ CFLAGS	:=	-g -Wall -Wdouble-promotion -std=gnu99 -O3 \
 			$(ARCH) $(DEFINES) `$(PREFIX)pkg-config --cflags SDL2_mixer`
 
 CFLAGS	+=	$(INCLUDE) -DARM11 -D__SWITCH__ -DGLQUAKE -DPLATFORM_DIRECTORY="nx" -DPLATFORM_RENDERER="gl" \
-			-DMAX_AI_COUNT=24 -DPLATFORM_CONFIRM_FLIPPED +DPLATFORM_KEYBOARD_SYSTEM
+			-DMAX_AI_COUNT=24 -DPLATFORM_CONFIRM_FLIPPED -DPLATFORM_KEYBOARD_SYSTEM \
+			-DPLATFORM_INPUT_GAMEPAD
 
 ifeq ($(WERROR),1)
 CFLAGS 		+= -Werror
@@ -129,6 +130,7 @@ COMMON_OBJS =	system.o \
 				images.c \
 				platform/nx/snd_nx.c \
 				input.c \
+				platform/nx/in_nx.c \
 				platform/nx/music_sdl.c \
 				snd_music.c \
 				platform/nx/gl/gl.c \
@@ -144,7 +146,7 @@ COMMON_OBJS =	system.o \
 				platform/nx/gl/gl_rmain.c \
 				platform/nx/gl/gl_rmisc.c \
 				platform/nx/gl/gl_rsurf.c \
-				platform/nx/gl/gl_screen.c \
+				render/r_screen.c \
 				platform/nx/gl/gl_warp.c \
 				platform/nx/gl/gl_vidnx.c \
 				platform/nx/r_part.c \
