+++ source/platform/psp/input.cpp
@@ -74,7 +74,7 @@ extern cvar_t in_mlook; //Heffo - mlook cvar
 
 extern bool system_has_right_stick;
 
-void IN_Init (void)
+void IN_PlatformInit(void)
 {
 	// Set up the controller.
 	sceCtrlSetSamplingCycle(0);
@@ -102,12 +102,12 @@ void IN_Init (void)
 	memcpy_vfpu(buttonToMessageKeyMap, buttonToConsoleKeyMap, sizeof(ButtonToKeyMap));
 }
 
-void IN_Shutdown (void)
+void IN_PlatformShutdown(void)
 {
 
 }
 
-void IN_Commands (void)
+void IN_PlatformCommands(void)
 {
 	// Changed in or out of key binding mode?
 	if ((bind_grab != 0) != readyToBindKeys)
@@ -247,140 +247,23 @@ void IN_Commands (void)
 	lastKeyMap = buttonToKeyMap;
 }
 
-float IN_CalcInput(int axis, float speed, float tolerance, float acceleration) {
 
-	float value = ((float) axis / 128.0f) - 1.0f;
-
-	if (value == 0.0f) {
-		return 0.0f;
-	}
-
-	float abs_value = fabs(value);
-
-	if (abs_value < tolerance) {
-		return 0.0f;
-	}
-
-	abs_value -= tolerance;
-	abs_value /= (1.0f - tolerance);
-	abs_value = powf(abs_value, acceleration);
-	abs_value *= speed;
-
-	if (value < 0.0f) {
-		value = -abs_value;
-	} else {
-		value = abs_value;
-	}
-	return value;
-}
-
-extern cvar_t scr_fov;
-extern int original_fov, final_fov;
-void IN_Move (usercmd_t *cmd)
+void IN_GetAnalogStick(in_analog_stick_id_t stick, in_analog_stick_t *value)
 {
-	V_StopPitchDrift();
-
-	// Read the pad state.
 	SceCtrlData pad;
 	sceCtrlPeekBufferPositive(&pad, 1);
-
-	// Convert the inputs to floats in the range [-1, 1].
-	// Implement the dead zone.
-	float speed;
-	float deadZone = in_tolerance.value;
-	float acceleration = in_acceleration.value;
-	float look_x, look_y;
-
-	//
-	// Analog look tweaks
-	//
-	speed = in_sensitivity.value;
-
-	// cut look speed in half when facing enemy, unless mag is empty
-	if ((in_aimassist.value) && (sv_player->v.facingenemy == 1) && cl.stats[STAT_CURRENTMAG] > 0) {
-		speed *= 0.5f;
-	}
-
-	// additionally, slice look speed when ADS/scopes
-	if (cl.stats[STAT_ZOOM] == 1)
-		speed *= 0.5f;
-	else if (cl.stats[STAT_ZOOM] == 2)
-		speed *= 0.25f;
-	
-	// Are we using the left or right stick for looking?
-	if (!in_anub_mode.value) { // Left
-		look_x = IN_CalcInput(pad.Lx, speed, deadZone, acceleration);
-		look_y = IN_CalcInput(pad.Ly, speed, deadZone, acceleration) * -1;
-	} else { // Right
-		look_x = IN_CalcInput(pad.Rsrv[0], speed, deadZone, acceleration);
-		look_y = IN_CalcInput(pad.Rsrv[1], speed, deadZone, acceleration) * -1;
-
-		if (!system_has_right_stick) {
-			look_x = look_y = 0;
-		}
-	}
-
-	const float yawScale = 30.0f;
-	cl.viewangles[YAW] -= yawScale * look_x * (float)host_frametime;
-
-	// Set the pitch.
-	const bool invertPitch = m_pitch.value > 0;
-	const float pitchScale = yawScale * (invertPitch ? 1 : -1);
-
-	cl.viewangles[PITCH] += pitchScale * look_y * (float)host_frametime;
-
-	// Don't look too far up or down.
-	if (cl.viewangles[PITCH] > 80.0f)
-		cl.viewangles[PITCH] = 80.0f;
-	if (cl.viewangles[PITCH] < -70.0f)
-		cl.viewangles[PITCH] = -70.0f;
-
-	// Ability to move with the left nub on NEW model systems
-	float move_x, move_y;
-	float input_x, input_y;
-
-	if (in_anub_mode.value) {
-		input_x = pad.Lx;
-		input_y = 255 - pad.Ly;
+	if (stick == IN_STICK_LEFT) {
+		value->x = ((float)pad.Lx - 127.5f) / 127.5f;
+		value->y = (127.5f - (float)pad.Ly) / 127.5f;
+	} else if (system_has_right_stick) {
+		value->x = ((float)pad.Rsrv[0] - 127.5f) / 127.5f;
+		value->y = (127.5f - (float)pad.Rsrv[1]) / 127.5f;
 	} else {
-		input_x = pad.Rsrv[0];
-		input_y = 255 - pad.Rsrv[1];
-
-		if (!system_has_right_stick) {
-			input_x = input_y = 128;
-		}
+		value->x = value->y = 0.0f;
 	}
+}
 
-	cl_backspeed = cl_forwardspeed = cl_sidespeed = sv_player->v.maxspeed;
-	cl_sidespeed *= 0.8f;
-	cl_backspeed *= 0.7f;
-
-	move_x = IN_CalcInput(input_x, cl_sidespeed, deadZone, acceleration);
-
-	if (input_y > 0)
-		move_y = IN_CalcInput(input_y, cl_forwardspeed, deadZone, acceleration);
-	else
-		move_y = IN_CalcInput(input_y, cl_backspeed, deadZone, acceleration);
-
-	// cypress -- explicitly setting instead of adding so we always prioritize
-	// analog movement over standard bindings if both are at play
-	if (move_x != 0 || move_y != 0) {
-		cmd->sidemove = move_x;
-		cmd->forwardmove = move_y;
-	} 
-
-	// crosshair stuff
-	if (cmd->forwardmove == 0.0 && cmd->sidemove == 0.0 && cl.onground) {
-		croshhairmoving = false;
-
-		crosshair_opacity += 22;
-
-		if (crosshair_opacity >= 255)
-			crosshair_opacity = 255;
-	} else {
-		croshhairmoving = true;
-		crosshair_opacity -= 8;
-		if (crosshair_opacity <= 128)
-			crosshair_opacity = 128;
-	}
+void IN_PlatformMove(usercmd_t *cmd)
+{
+	(void)cmd;
 }
