+++ source/platform/psp2/in_psp2.c
@@ -37,8 +37,6 @@ extern void Log (const char *format, ...);
 
 #define lerp(value, from_max, to_max) ((((value*10) * (to_max*10))/(from_max*10))/10)
 
-float IN_CalcInput(int axis, float speed, float tolerance, float acceleration);
-
 extern bool croshhairmoving;
 extern float crosshair_opacity;
 
@@ -46,8 +44,9 @@ uint64_t rumble_tick = 0;
 SceCtrlData oldanalogs, analogs;
 SceMotionState motionstate;
 
-void IN_Init (void)
+void IN_PlatformInit(void)
 {
+  Cvar_SetValue("in_anub_mode", 1);
   Cvar_RegisterVariable (&m_filter);
   Cvar_RegisterVariable (&retrotouch);
   Cvar_RegisterVariable (&pstv_rumble);
@@ -67,11 +66,11 @@ void IN_Init (void)
   sceMotionStartSampling();
 }
 
-void IN_Shutdown (void)
+void IN_PlatformShutdown(void)
 {
 }
 
-void IN_Commands (void)
+void IN_PlatformCommands(void)
 {
 }
 
@@ -114,163 +113,21 @@ void IN_StopRumble (void)
 // 	}
 // }
 
-void IN_Move (usercmd_t *cmd)
+
+void IN_GetAnalogStick(in_analog_stick_id_t stick, in_analog_stick_t *value)
 {
-	// ANALOGS
-	sceCtrlPeekBufferPositive(0, &analogs, 1);
-	int left_x = analogs.lx - 127;
-	int left_y = analogs.ly - 127;
-	int right_x = analogs.rx - 127;
-	int right_y = analogs.ry - 127;
-
-	// Convert the inputs to floats in the range [-1, 1].
-	// Implement the dead zone.
-	float speed;
-	float deadZone = in_tolerance.value;
-	float acceleration = in_acceleration.value;
-	float look_x, look_y;
-
-	// 
-	// Analog look tweaks
-	//
-	speed = sensitivity.value;
-
-	// cut look speed in half when facing enemy, unless mag is empty
-	if ((in_aimassist.value) && (sv_player->v.facingenemy == 1) && cl.stats[STAT_CURRENTMAG] > 0) {
-		speed *= 0.5f;
-	}
-	// additionally, slice look speed when ADS/scopes
-	if (cl.stats[STAT_ZOOM] == 1)
-		speed *= 0.5f;
-	else if (cl.stats[STAT_ZOOM] == 2)
-		speed *= 0.25f;
-	
-	look_x = IN_CalcInput(right_x, speed, deadZone, acceleration);
-	look_y = IN_CalcInput(right_y, speed, deadZone, acceleration);
-
-	const float yawScale = 30.0f;
-	cl.viewangles[YAW] -= yawScale * look_x * (float)host_frametime;
-
-	// Set the pitch.
-	const bool invertPitch = m_pitch.value > 0;
-	const float pitchScale = yawScale * (invertPitch ? 1 : -1);
-
-	cl.viewangles[PITCH] += pitchScale * look_y * (float)host_frametime;
-
-	// Don't look too far up or down.
-	if (cl.viewangles[PITCH] > 80.0f)
-		cl.viewangles[PITCH] = 80.0f;
-	if (cl.viewangles[PITCH] < -70.0f)
-		cl.viewangles[PITCH] = -70.0f;
-
-	// Ability to move with the left nub on NEW model systems
-	float move_x, move_y;
-
-	cl_backspeed = cl_forwardspeed = cl_sidespeed = sv_player->v.maxspeed;
-	cl_sidespeed *= 0.8f;
-	cl_backspeed *= 0.7f;
-
-	move_x = IN_CalcInput(left_x, cl_sidespeed, deadZone, acceleration);
-
-	if (left_y > 0)
-		move_y = IN_CalcInput(left_y, cl_forwardspeed, deadZone, acceleration);
-	else
-		move_y = IN_CalcInput(left_y, cl_backspeed, deadZone, acceleration);
-
-	// cypress -- explicitly setting instead of adding so we always prioritize
-	// analog movement over standard bindings if both are at play
-	if (move_x != 0 || move_y != 0) {
-		cmd->sidemove = move_x;
-		cmd->forwardmove = move_y;
-	} 
-
-	// crosshair stuff
-	if (move_x < 50 && move_x > -50 && move_y < 50 && move_y > -50) {
-		croshhairmoving = false;
-
-		crosshair_opacity += 22;
-
-		if (crosshair_opacity >= 255)
-			crosshair_opacity = 255;
+	SceCtrlData state;
+	sceCtrlPeekBufferPositive(0, &state, 1);
+	if (stick == IN_STICK_LEFT) {
+		value->x = ((float)state.lx - 127.5f) / 127.5f;
+		value->y = (127.5f - (float)state.ly) / 127.5f;
 	} else {
-		croshhairmoving = true;
-		crosshair_opacity -= 8;
-		if (crosshair_opacity <= 128)
-			crosshair_opacity = 128;
+		value->x = ((float)state.rx - 127.5f) / 127.5f;
+		value->y = ((float)state.ry - 127.5f) / 127.5f;
 	}
+}
 
-	/*
-	// Left analog support for player movement
-	float x_mov = abs(left_x) < 30 ? 0 : (left_x * cl_sidespeed) * 0.01;
-	float y_mov = abs(left_y) < 30 ? 0 : (left_y * (left_y > 0 ? cl_backspeed : cl_forwardspeed)) * 0.01;
-	cmd->forwardmove -= y_mov;
-	//if (gl_xflip.value) cmd->sidemove -= x_mov;
-	cmd->sidemove += x_mov;
-
-	// Right analog support for camera movement
-	//IN_RescaleAnalog(&right_x, &right_y, 30);
-	float x_cam = (right_x * sensitivity.value) * 0.008;
-	float y_cam = (right_y * sensitivity.value) * 0.008;
-	//if (gl_xflip.value) cl.viewangles[YAW] += x_cam;
-	cl.viewangles[YAW] -= x_cam;
-	V_StopPitchDrift();
-	cl.viewangles[PITCH] += y_cam;
-	*/
-	// TOUCH SUPPORT
-
-	// Retrotouch support for camera movement
-	/*
-	SceTouchData touch;
-	if (retrotouch.value){
-		sceTouchPeek(SCE_TOUCH_PORT_BACK, &touch, 1);
-		if (touch.reportNum > 0) {
-			int raw_x = lerp(touch.report[0].x, 1919, 960);
-			int raw_y = lerp(touch.report[0].y, 1087, 544);
-			int touch_x = raw_x - 480;
-			int touch_y = raw_y - 272;
-			x_cam = abs(touch_x) < 20 ? 0 : touch_x * psvita_back_sensitivity_x.value * 0.008;
-			y_cam = abs(touch_y) < 20 ? 0 : touch_y * psvita_back_sensitivity_x.value * 0.008;
-			cl.viewangles[YAW] -= x_cam;
-			V_StopPitchDrift();
-			cl.viewangles[PITCH] += y_cam;
-		}
-	}
-
-	if (psvita_touchmode.value == 1)
-	{
-		sceTouchPeek(SCE_TOUCH_PORT_FRONT, &touch, 1);
-		if (touch.reportNum > 0) {
-			int raw_x = lerp(touch.report[0].x, 1919, 960);
-			int raw_y = lerp(touch.report[0].y, 1087, 544);
-			int touch_x = raw_x - 480;
-			int touch_y = raw_y - 272;
-			x_cam = abs(touch_x) < 20 ? 0 : touch_x * psvita_front_sensitivity_x.value * 0.008;
-			y_cam = abs(touch_y) < 20 ? 0 : touch_y * psvita_front_sensitivity_y.value * 0.008;
-			cl.viewangles[YAW] -= x_cam;
-			V_StopPitchDrift();
-			cl.viewangles[PITCH] += y_cam;
-		}
-	}
-
-  // gyro analog support for camera movement
-
-  if (motioncam.value){
-    sceMotionGetState(&motionstate);
-
-    // not sure why YAW or the horizontal x axis is the controlled by angularVelocity.y
-    // and the PITCH or the vertical y axis is controlled by angularVelocity.x but its what seems to work
-    float x_gyro_cam = motionstate.angularVelocity.y * motion_horizontal_sensitivity.value;
-    float y_gyro_cam = motionstate.angularVelocity.x * motion_vertical_sensitivity.value;
-
-    //if (gl_xflip.value)
-      //cl.viewangles[YAW] -= x_gyro_cam;
-    //else
-    cl.viewangles[YAW] += x_gyro_cam;
-
-    V_StopPitchDrift();
-
-      cl.viewangles[PITCH] -= y_gyro_cam;
-  }
-	*/
-	//cl.viewangles[PITCH] = COM_Clamp(cl.viewangles[PITCH], -70, 80);
+void IN_PlatformMove(usercmd_t *cmd)
+{
+	(void)cmd;
 }
